vol: Repair _VHandleVolumeHeader retry logic
Commit 2cd1335257f0 "Abstract /vicepX header traversal" introduced
routines VWalkVolumeHeaders() and _VHandleVolumeHeader(). In
_VHandleVolumeHeader(), if 'locked' is 0 and we encounter an error, we
are supposed to lock the partition, and retry reading the volume header.
However, this retry logic is never actually run, due to a few bugs:

First, VWalkVolumeHeaders() calls _VHandleVolumeHeader() with 'locked' set
to -1 instead of 0. So even though the partition lock isn't held,
_VHandleVolumeHeader() never retries reading the header, since 'locked'
is nonzero.

Another bug is a typo in _VHandleVolumeHeader(): '#ifdef
AFSFS_DEMAND_ATTACH_FS' will never be true, since the relevant define is
actually AFS_DEMAND_ATTACH_FS. So we never try to unlock the partition
if 'locked' is set. This is actually fortunate, because the caller
currently always mistakenly passes -1 for 'locked' even though the
partition lock isn't held.

Finally, the call to VPartHeaderUnlock() is missing its second argument
'locktype'. This was never noticed because of the AFS_DEMAND_ATTACH_FS
typo, and so this call is never currently compiled.

To fix all of these: change VWalkVolumeHeaders() to call
_VHandleVolumeHeader() with 'locked' set to 0, fix the
AFS_DEMAND_ATTACH_FS typo, and call VPartHeaderUnlock() with WRITE_LOCK
to match the VPartHeaderLock(WRITE_LOCK) call in VWalkVolumeHeaders().

Change-Id: Ic998c0e32211bea9466af10ad3be20ef4f9221be
Reviewed-on: https://gerrit.openafs.org/16581
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>