vol: Set salvage linktable linkcount to nVols
Since OpenAFS 1.0, salvager has been able to recreate a missing link
table.  However, the created link table is initialized (by icreate) with
a linkcount value of 1 for itself.  This is only correct if there is one
volume in the volume group.

If a volume group contains more than one volume, and the salvager must
recreate the linktable, then after salvaging, the generated linktable
will have a linkcount that is too low. If the other volumes in the
volume group are deleted, the linktable will be prematurely deleted,
probably triggering another salvage, recreating the linktable again.

To avoid all this, modify CreateLinkTable to set the linktable's own
link count (inode 0) to the number of volumes in the volume group.

Change-Id: I319641a2ccd812ade3bf509c3c07c054a6d057a9
Reviewed-on: https://gerrit.openafs.org/16586
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>