volser: Introduce GetLockedEntry
In preparation for a future commit, consolidate common VL_SetLock and
VL_GetEntryById logic into a new routine GetLockedEntry.

This commit should incur no practical change in behavior, but there is a
minor internal behavior change for one case:

For the case in UV_ConvertRO(), previously we called VL_SetLock with the
RW volume id (with volume type RWVOL), and called VL_GetEntryByID with
the RO volume id (with the volume type -1); now we call both with the RW
volume id. We get these volume IDs from the same VL entry, and they must
refer to the same VL entry in order for this function to work properly,
so this difference should not matter.

Change-Id: I76faf3d184c397bceff27ba3a857b5c80c88e814
Reviewed-on: https://gerrit.openafs.org/14350
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>