volser: Always fetch locked entry in CheckVolume
CheckVolume accepts an optional nvldbentry as an argument.  When this
argument is NULL, CheckVolume correctly locks the nvldbentry before
obtaining the entry and making changes.  However, if an nvldbentry is
passed, CheckVolume merely copies it to a work entry; it also neglects
to obtain the current state of the entry after locking it for
modification.

Change CheckVolume to obtain a locked copy of the nvldbentry if
modifications are required.

Change-Id: Ie98bc57af1cc2955d10982f44abd4c55755c6244
Reviewed-on: https://gerrit.openafs.org/14356
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>